[DH-813] Move IsAssociatedToInvestmentProjectFilter
